scene controller methods

+++ UserInterface/Panels/FigurePanel.cs
@@ -47,11 +47,11 @@ namespace Render3D.UserInterface.Panels
             txtFigureRadius.Text = "";
         }
 
-        private decimal TryToParse(string figureRadiusString)
+        private double TryToParse(string figureRadiusString)
         {
             try
             {
-               decimal radius= Decimal.Parse(figureRadiusString);
+               double radius= double.Parse(figureRadiusString);
                 return radius; 
             }catch 
             {
